# Use stackalloc char&byte arrays for decoding Url values in HttpUtiliy.ParseQueryString