--- a/src/libraries/System.Web.HttpUtility/src/System/Web/Util/HttpEncoder.cs
+++ b/src/libraries/System.Web.HttpUtility/src/System/Web/Util/HttpEncoder.cs
@@ -258,7 +258,9 @@ namespace System.Web.Util
                 return null;
             }
 
-            UrlDecoder helper = new UrlDecoder(count, encoding);
+            UrlDecoder helper = count < 256
+                ? new UrlDecoder(stackalloc char[count], stackalloc byte[count], encoding)
+                : new UrlDecoder(new char[count], new byte[count], encoding);
 
             // go through the bytes collapsing %XX and %uXXXX and appending
             // each byte as byte, with exception of %uXXXX constructs that
@@ -321,8 +323,20 @@ namespace System.Web.Util
                 return null;
             }
 
+            return UrlDecode(value.AsSpan(), encoding);
+        }
+
+        internal static string UrlDecode(ReadOnlySpan<char> value, Encoding encoding)
+        {
+            if (value.IsEmpty)
+            {
+                return string.Empty;
+            }
+
             int count = value.Length;
-            UrlDecoder helper = new UrlDecoder(count, encoding);
+            UrlDecoder helper = count < 256
+                ? new UrlDecoder(stackalloc char[count], stackalloc byte[count], encoding)
+                : new UrlDecoder(new char[count], new byte[count], encoding);
 
             // go through the string's chars collapsing %XX and %uXXXX and
             // appending each char as char, with exception of %XX constructs
@@ -626,17 +640,15 @@ namespace System.Web.Util
         }
 
         // Internal class to facilitate URL decoding -- keeps char buffer and byte buffer, allows appending of either chars or bytes
-        private sealed class UrlDecoder
+        private ref struct UrlDecoder
         {
-            private readonly int _bufferSize;
-
             // Accumulate characters in a special array
             private int _numChars;
-            private readonly char[] _charBuffer;
+            private readonly Span<char> _charBuffer;
 
             // Accumulate bytes for decoding into characters in a special array
             private int _numBytes;
-            private byte[]? _byteBuffer;
+            private readonly Span<byte> _byteBuffer;
 
             // Encoding to convert chars to bytes
             private readonly Encoding _encoding;
@@ -645,19 +657,17 @@ namespace System.Web.Util
             {
                 if (_numBytes > 0)
                 {
-                    Debug.Assert(_byteBuffer != null);
-                    _numChars += _encoding.GetChars(_byteBuffer, 0, _numBytes, _charBuffer, _numChars);
+                    Debug.Assert(!_byteBuffer.IsEmpty);
+                    _numChars += _encoding.GetChars(_byteBuffer.Slice(0, _numBytes), _charBuffer.Slice(_numChars));
                     _numBytes = 0;
                 }
             }
 
-            internal UrlDecoder(int bufferSize, Encoding encoding)
+            internal UrlDecoder(Span<char> charBuffer, Span<byte> byteBuffer, Encoding encoding)
             {
-                _bufferSize = bufferSize;
+                _charBuffer = charBuffer;
+                _byteBuffer = byteBuffer;
                 _encoding = encoding;
-
-                _charBuffer = new char[bufferSize];
-                // byte buffer created on demand
             }
 
             internal void AddChar(char ch)
@@ -681,8 +691,6 @@ namespace System.Web.Util
                                 else
                 */
                 {
-                    _byteBuffer ??= new byte[_bufferSize];
-
                     _byteBuffer[_numBytes++] = b;
                 }
             }
@@ -694,7 +702,7 @@ namespace System.Web.Util
                     FlushBytes();
                 }
 
-                return _numChars > 0 ? new string(_charBuffer, 0, _numChars) : "";
+                return _numChars > 0 ? _charBuffer.Slice(0, _numChars).ToString() : "";
             }
         }
     }