1. Convert UrlDecoder to ref struct
2. If Url is less than 256 characters, do no allocate char&byte arrays
3. Replace substring operation with Span