@TrayanZapryanov 2024-05-28
Benchmark:
``` C#
[MemoryDiagnoser]
public class HttpUtilityBenchmarks
{
	[Benchmark(Baseline = true)]
	public NameValueCollection ParseQueryString() => HttpUtility.ParseQueryString("https://portal.azure.com/#@testgroup.onmicrosoft.com/dashboard/arm/subscriptions/111111-abac-4067-81ba-5ae054d891b5/" +
	                                                                                "resourcegroups/dashboards/providers/microsoft.portal/dashboards/3456268-8630-60f7-aaac-eca38aca231");

	[Benchmark]
	public NameValueCollection ParseQueryString_Mix() => HttpUtility.ParseQueryString("ReturnUrl=http%3a%2f%2flocalhost%2flogin%2fauthenticate%3fReturnUrl%3dhttp%3a%2f%2flocalhost%2fsecured_area%26__provider__%3dgoogle");
}
```

Results:
Machine:
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3593/23H2/2023Update/SunValley3)
11th Gen Intel Core i9-11900K 3.50GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.100-preview.4.24267.66
  [Host]     : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI

Before:
| Method               | Mean     | Error   | StdDev  | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|--------------------- |---------:|--------:|--------:|------:|--------:|-------:|-------:|----------:|------------:|
| ParseQueryString     | 518.0 ns | 9.81 ns | 9.18 ns |  1.00 |    0.00 | 0.1955 |      - |    1.6 KB |        1.00 |
| ParseQueryString_Mix | 430.0 ns | 8.37 ns | 8.59 ns |  0.83 |    0.02 | 0.1931 | 0.0005 |   1.58 KB |        0.99 |

PR:
| Method               | Mean     | Error   | StdDev  | Ratio | Gen0   | Allocated | Alloc Ratio |
|--------------------- |---------:|--------:|--------:|------:|-------:|----------:|------------:|
| ParseQueryString     | 490.7 ns | 4.03 ns | 3.77 ns |  1.00 | 0.1059 |     888 B |        1.00 |
| ParseQueryString_Mix | 337.3 ns | 4.27 ns | 3.78 ns |  0.69 | 0.0820 |     688 B |        0.77 |